Add testing data for the todo tests
There are two design decisions being made here.

First, we're factoring out the testing data into its own file. (Unlike the user tests, which keep a copy of the testing data in each .spec.ts file.) This prevents some code duplication and gives us more flexibility for the future. (For example, it makes it easier to add multiple sets of test data if we wanted.)

Second, we're storing the testing data in an exported variable in a TypeScript file, rather than as a JSON file. This makes it slightly easier to import the test data.