Add empty files for the todo-table component
This component will display the todos in a Material table.

Material tables are nice because they can display a lot of data in a compact format. Also, they make sorting a breeze and a half—you just have to attach a mat-sort-header and you get automatic sorting by arbitrary columns with a convenient and idiomatic UI.