Make the tests of the limiter pass
The problem was that, if the limiter was not selected to anything—eg, if it was on the default—then the array limitedTodos would never be assigned and it would be undefined.

The fix for this is to put an else clause: if the limiter is not selected to anything, we just assign limitedTodos to filteredTodos.

Closes #6.