Use a method call as the table's data source rather than a field
Using a method call lets us keep the data source 'live': whenever the underlying `todo` (which we're getting from @Input) changes, the method will see that change. (But an array stored in a field wouldn't have seen that change.)

Now, we do still have a problem, which is that, when the sort header is clicked, the table doesn't know to re-render its rows. (If the @Input binding changes, the table is smart enough to see that and re-render. But changes to the sort header are orthogonal to changes in the @Input binding, and so we need to explicitly let the table know when that happens. This will be resolved in a future commit.)